feat: 코인상세 페이지가 생성되고, 차트뷰컨트롤러를 생성할 때 캔들스틱 API를 호출하도록 repository 위치 변경
- ChartView에서 담당하던 API 호출-응답을 ChartByTimesViewController의 책임으로 변경
