+++ CryptocurrencyExchange/CryptocurrencyExchange/ChartsByCoin/ChartByTimes/Chart/ViewModel/ChartViewModel.swift
@@ -10,13 +10,10 @@ import Charts
 
 class ChartViewModel {
     var dataEntries: Observable<CandleChartDataEntries?>
-    var repository: CandleStickChartRepository
 
-    init(dataEntries: Observable<CandleChartDataEntries?> = Observable(nil),
-         repository: CandleStickChartRepository
+    init(dataEntries: Observable<CandleChartDataEntries?> = Observable(nil)
     ) {
         self.dataEntries = dataEntries
-        self.repository = repository
     }
     
     /// 차트를 보여줄 때 원하는 갯수만큼 보이도록 zoomFactor 계산
@@ -29,14 +26,5 @@ class ChartViewModel {
             xValue: xValue,
             yValue: 0.0
         )
-    }
-    
-    /// 레파지토리에 candleStick API 호출하도록 요청
-    func getCandleStickData(parameter: CandleStickParameters) {
-        self.repository.getCandleStickData(
-            parameter: parameter
-        ) { stickValues in
-                self.dataEntries.value = try? CandleChartDataEntries(with: stickValues)
-            }
     }
 }
