+++ CryptocurrencyExchange/CryptocurrencyExchange/ChartsByCoin/ChartByTimes/ChartByTimesViewModel.swift
@@ -9,8 +9,23 @@ import Foundation
 
 class ChartByTimesViewModel: XIBInformation {
     var nibName: String?
+    var repository: CandleStickChartRepository
+
     
-    init(nibName: String?) {
+    init(
+        nibName: String?,
+        repository: CandleStickChartRepository
+    ) {
         self.nibName = nibName
+        self.repository = repository
+    }
+    
+    /// 레파지토리에 candleStick API 호출하도록 요청
+    func getCandleStickData(parameter: CandleStickParameters) {
+        self.repository.getCandleStickData(
+            parameter: parameter
+        ) { stickValues in
+            // 추후 db에 저장해야 함
+        }
     }
 }
